rocky/curs_manager/viewsets.py
from rest_framework.viewsets import ModelViewSet

from .models import Student
from .serializers import StudentSerializer

class StudentViewSet(ModelViewSet):
    queryset = Student.objects.all()
    serializer_class = StudentSerializer