requirements.txt
django==4.1.4
psycopg2==2.9.5
django-debug-toolbar==3.8.1
django-grappelli==3.0.4
djangorestframework==3.14.0